refactor: 필요 없는 도메인 제거 및 정적 팩터리 메서드로 객체 생성할 수 있게 리팩토링 진행
